SONAR-4898 - Fixed AppTest (NPE when App is not started)

+++ sonar-application/src/main/java/org/sonar/application/App.java
@@ -124,8 +124,12 @@ public class App implements ProcessMXBean {
       logger.info("Shutting down server");
       monitor.interrupt();
       monitor = null;
-      elasticsearch.terminate();
-      server.terminate();
+      if (elasticsearch != null) {
+        elasticsearch.terminate();
+      }
+      if (server != null) {
+        server.terminate();
+      }
     }
   }
